add useGuideScroll hook
