+++ template/web/app/buy-me-coffee/_components/Guide.tsx
@@ -1,6 +1,6 @@
-import { useEffect } from 'react';
 import { usePathname } from 'next/navigation';
 import CodeBlock from '@/components/code-block/CodeBlock';
+import useGuideScroll from '@/components/layout/guide/useGuideScroll';
 
 const codeStep1 = `\`\`\`solidity
 if (msg.value < price * numCoffees) {
@@ -69,38 +69,7 @@ return slice;
 export default function Guide() {
   const pathname = usePathname();
 
-  useEffect(() => {
-    function convertRemToPixels(rem: number) {
-      return rem * parseFloat(getComputedStyle(document.documentElement).fontSize);
-    }
-
-    function handleScroll() {
-      const pageYOffset = window.scrollY;
-      let newActiveSectionId = null;
-
-      window.document.querySelectorAll('section[id]').forEach((section) => {
-        const sectionOffsetTop = (section as HTMLElement).offsetTop;
-
-        if (pageYOffset >= sectionOffsetTop - convertRemToPixels(7)) {
-          newActiveSectionId = section.id;
-        }
-      });
-
-      window.document
-        .querySelectorAll(`aside nav li a[href]`)
-        .forEach((linkItem) => linkItem.setAttribute('data-active', 'false'));
-
-      window.document
-        .querySelector(`nav li a[href="#${newActiveSectionId}"]`)
-        ?.setAttribute('data-active', 'true');
-    }
-
-    window.addEventListener('scroll', handleScroll);
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, []);
+  useGuideScroll();
 
   return (
     <>
