+++ template/web/app/home/_components/Guide.tsx
@@ -1,6 +1,6 @@
-import { useEffect } from 'react';
 import { usePathname } from 'next/navigation';
 import CodeBlock from '@/components/code-block/CodeBlock';
+import useGuideScroll from '@/components/layout/guide/useGuideScroll';
 
 const codeStep1 = `\`\`\`bash
 $ npx @coinbase/build-onchain-apps@latest create`;
@@ -91,38 +91,7 @@ const tocContent = [
 ];
 
 export default function Guide() {
-  useEffect(() => {
-    function convertRemToPixels(rem: number) {
-      return rem * parseFloat(getComputedStyle(document.documentElement).fontSize);
-    }
-
-    function handleScroll() {
-      const pageYOffset = window.scrollY;
-      let newActiveSectionId = null;
-
-      window.document.querySelectorAll('section[id]').forEach((section) => {
-        const sectionOffsetTop = (section as HTMLElement).offsetTop;
-
-        if (pageYOffset >= sectionOffsetTop - convertRemToPixels(7)) {
-          newActiveSectionId = section.id;
-        }
-      });
-
-      window.document
-        .querySelectorAll(`aside nav li a[href]`)
-        .forEach((linkItem) => linkItem.setAttribute('data-active', 'false'));
-
-      window.document
-        .querySelector(`nav li a[href="#${newActiveSectionId}"]`)
-        ?.setAttribute('data-active', 'true');
-    }
-
-    window.addEventListener('scroll', handleScroll);
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, []);
+  useGuideScroll();
 
   return (
     <>
