+++ template/web/src/components/layout/guide/useGuideScroll.ts
@@ -1,0 +1,36 @@
+import { useEffect } from 'react';
+
+export default function useGuideScroll() {
+  useEffect(() => {
+    function convertRemToPixels(rem: number) {
+      return rem * parseFloat(getComputedStyle(document.documentElement).fontSize);
+    }
+
+    function handleScroll() {
+      const pageYOffset = window.scrollY;
+      let newActiveSectionId = null;
+
+      window.document.querySelectorAll('section[id]').forEach((section) => {
+        const sectionOffsetTop = (section as HTMLElement).offsetTop;
+
+        if (pageYOffset >= sectionOffsetTop - convertRemToPixels(7)) {
+          newActiveSectionId = section.id;
+        }
+      });
+
+      window.document
+        .querySelectorAll(`aside nav li a[href]`)
+        .forEach((linkItem) => linkItem.setAttribute('data-active', 'false'));
+
+      window.document
+        .querySelector(`nav li a[href="#${newActiveSectionId}"]`)
+        ?.setAttribute('data-active', 'true');
+    }
+
+    window.addEventListener('scroll', handleScroll);
+
+    return () => {
+      window.removeEventListener('scroll', handleScroll);
+    };
+  }, []);
+}
